pyproject.toml
[project]
name = "boggart-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Chandler Berry", email = "chandler.r.berry@proton.me" }
]
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.55.0",
    "discord-py>=2.5.2",
    "httpx>=0.28.1",
    "openai>=1.91.0",
    "pydantic>=2.11.7",
    "pydantic-ai-slim[anthropic,duckduckgo,mcp,openai]>=0.3.4",
    "pydantic-settings>=2.10.1",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "ruff>=0.12.0",
]

[project.scripts]
boggart = "boggart_2.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.format]
quote-style = "single"

[tool.pyright]
venvPath='.'
venv=".venv"
exclude=['.venv']
